requirements.txt
pyaml
aiohttp[speedups]
aiofile
websockets
